+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/BaseImageParameters.java
@@ -35,8 +35,8 @@ import org.gradle.api.tasks.Optional;
 public class BaseImageParameters {
 
   private final AuthParameters auth;
-  private Property<String> image;
-  @Nullable private String credHelper;
+  private final Property<String> image;
+  private final CredHelperParameters credHelper;
   private final PlatformParametersSpec platformParametersSpec;
   private final ListProperty<PlatformParameters> platforms;
 
@@ -46,6 +46,8 @@ public class BaseImageParameters {
     platforms = objectFactory.listProperty(PlatformParameters.class);
     image = objectFactory.property(String.class);
     platformParametersSpec = objectFactory.newInstance(PlatformParametersSpec.class, platforms);
+    credHelper =
+        objectFactory.newInstance(CredHelperParameters.class, PropertyNames.FROM_CRED_HELPER);
 
     PlatformParameters amd64Linux = new PlatformParameters();
     amd64Linux.setArchitecture("amd64");
@@ -92,18 +94,18 @@ public class BaseImageParameters {
     this.image.set(image);
   }
 
-  @Input
-  @Nullable
+  @Nested
   @Optional
-  public String getCredHelper() {
-    if (System.getProperty(PropertyNames.FROM_CRED_HELPER) != null) {
-      return System.getProperty(PropertyNames.FROM_CRED_HELPER);
-    }
+  public CredHelperParameters getCredHelper() {
     return credHelper;
   }
 
-  public void setCredHelper(String credHelper) {
-    this.credHelper = credHelper;
+  public void setCredHelper(String helper) {
+    this.credHelper.setHelper(helper);
+  }
+
+  public void credHelper(Action<? super CredHelperParameters> action) {
+    action.execute(credHelper);
   }
 
   @Nested