+++ jib-maven-plugin/src/integration-test/java/com/google/cloud/tools/jib/maven/BuildDockerMojoIntegrationTest.java
@@ -261,4 +261,24 @@ public class BuildDockerMojoIntegrationTest {
           ex.getMessage(), CoreMatchers.containsString("but is required to be [,1.0]"));
     }
   }
+
+  @Test
+  public void testCredHelperConfigurationSimple()
+      throws DigestException, VerificationException, IOException, InterruptedException {
+    String targetImage = "simpleimage:maven" + System.nanoTime();
+    buildToDockerDaemon(simpleTestProject, targetImage, "pom-cred-helper-1.xml");
+    Assert.assertEquals(
+        "Hello, world. \n1970-01-01T00:00:01Z\n",
+        new Command("docker", "run", "--rm", "my-artifact-id:1").run());
+  }
+
+  @Test
+  public void testCredHelperConfigurationComplex()
+      throws DigestException, VerificationException, IOException, InterruptedException {
+    String targetImage = "simpleimage:maven" + System.nanoTime();
+    buildToDockerDaemon(simpleTestProject, targetImage, "pom-cred-helper-2.xml");
+    Assert.assertEquals(
+        "Hello, world. \n1970-01-01T00:00:01Z\n",
+        new Command("docker", "run", "--rm", "my-artifact-id:1").run());
+  }
 }