+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/MavenRawConfiguration.java
@@ -51,8 +51,8 @@ public class MavenRawConfiguration implements RawConfiguration {
   }
 
   @Override
-  public Optional<String> getFromCredHelper() {
-    return Optional.ofNullable(jibPluginConfiguration.getBaseImageCredentialHelperName());
+  public CredHelperConfiguration getFromCredHelper() {
+    return jibPluginConfiguration.getBaseImageCredHelperConfig();
   }
 
   @Override
@@ -66,8 +66,8 @@ public class MavenRawConfiguration implements RawConfiguration {
   }
 
   @Override
-  public Optional<String> getToCredHelper() {
-    return Optional.ofNullable(jibPluginConfiguration.getTargetImageCredentialHelperName());
+  public CredHelperConfiguration getToCredHelper() {
+    return jibPluginConfiguration.getTargetImageCredentialHelperConfig();
   }
 
   @Override